Make process.py robust for when not yet calibrated
process.py does not attempt to process the files, if calibration has not
been solved yet. Only if test.fits is available, the processing will
start. This makes process.py, just wait for new files, to try the
calibrate on. This is handy if you started process.py, while the sky is
still bright or clouded.